requirements.txt
Telethon~=1.16.4
PySocks==1.6.7
loguru